32-Artist Claims for auth of user actions

Related work items: #32

+++ backend/RS1_2024_25.API/Endpoints/AlbumEndpoints/AlbumDeleteEndpoint.cs
@@ -1,13 +1,18 @@
 ﻿using Azure.Core;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore;
 using RS1_2024_25.API.Data;
 using RS1_2024_25.API.Data.Models;
 using RS1_2024_25.API.Helper.Api;
+using RS1_2024_25.API.Services;
+using System.Security.Claims;
 
 namespace RS1_2024_25.API.Endpoints.AlbumEndpoints
 {
-    public class AlbumDeleteEndpoint(ApplicationDbContext db) : MyEndpointBaseAsync.WithRequest<int>.WithoutResult
+    public class AlbumDeleteEndpoint(ApplicationDbContext db, TokenProvider tp) : MyEndpointBaseAsync.WithRequest<int>.WithoutResult
     {
+        [Authorize]
         [HttpDelete("{id}")]
         public override async Task HandleAsync(int id, CancellationToken cancellationToken = default)
         {
@@ -18,7 +23,23 @@ namespace RS1_2024_25.API.Endpoints.AlbumEndpoints
                 throw new KeyNotFoundException("Album not found");
             }
 
-            db.Albums.Remove(a);
+            var jwt = tp.GetDecodedJwt(Request);
+            bool isAdmin = jwt.Claims.FirstOrDefault(c => c.Type == ClaimTypes.Role)!.Value == "Admin";
+            if (isAdmin)
+            {
+                db.Albums.Remove(a);
+            }
+            else
+            {
+                int userId = int.Parse(jwt.Subject);
+                UserArtist? uaCheck = await db.UserArtists.FirstOrDefaultAsync(ua => ua.MyAppUserId == userId && ua.ArtistId == a.ArtistId && ua.IsUserOwner);
+                if (uaCheck == null)
+                {                   
+                    throw new KeyNotFoundException("Artist not found");
+                }
+                db.Albums.Remove(a);
+            }
+
             await db.SaveChangesAsync();
         }
     }
